feature/resume/deleteResume : Add function to delete resume

+++ controllers/resumeController/delete.js
@@ -4,8 +4,12 @@ module.exports = async (req, res, next) => {
   const { resumeId } = req.body;
   const { _id } = req.user;
   try {
-    await Resume.deleteOne({ _id: resumeId, userId: _id });
-    res.status(200).json({ isDeleted: true });
+    if (req.user) {
+      await Resume.deleteOne({ _id: resumeId, userId: _id });
+      res.status(200).json({ isDeleted: true });
+    } else {
+      res.status(401).json({ isDeleted: false });
+    }
   } catch (err) {
     next(err);
   }
